day10p2: solve it!
The code comments are enlightening with respect to the algorithm. The
essential idea is to

- group by angle (equivalence classes of visibility) from the station;
- sort the groups by angle;
- sort each group by distance;
- peel heads off each group in order.

The complication arises from the grid orientation and laser sweep:

       -y (also laser start point)
       ^
       |θ/ (→ laser direction)
       |/
-x <---+---> x
       |
       |
       V
       y

But atan2 gives results in [-π, π], so I needed "-y" to be "-π", and so
on around the circle.

The simplest way to do this was to linearly transform the coordinate
system (x,y) → (y, -x) as in the comments. Then sorting gives the
correct angle order--almost. It turns out that in that mapping, the
start point (-y in the original system) gets mapped by atan2 to π! So a
quick check is in order for the start points relative to the other
points.

The check is simpler than it might otherwise be, since when comparing
elements of two classes to compare the classes, no two can be on the
same angle from the station. Therefore, no two can be on -y at the same
time!