copy files
For the new one, I think I need to test points in order (à la
Point.compare): if a point is in the beam, I can check if an NxN square
fits in the beam by checking only the 3 remaining corners.

I should be able to give the closest coordinate by ordering on Manhattan
distance, though if I check all the points in order, that should be the
"one" I started with.

I can probably do this recursively with a queue of points to check…
now, how to generate the "list of next points"?