day19p2: solve it!
A square is contained if the top left and bottom right points are
contained (and this is the "first best fit"). Then I need to grab the
points in the square and find the closest one. (I nearly forgot this for
a long time! Some assumptions are just wrong…)

The "order by Point.pi" method would probably have worked too, but it
seemed to be taking too long (π(1302,1155) = 3020808).

So, instead of checking points in order, I check the left-most edge of
the tractor beam. For each row, I can check every column until I find
the beam. The complicated piece of this, though, is that the first
handful of rows don't actually have anything in the beam! So we skip
that (see the calculation for firstX and firstY---if you need to skip
more than 100 points, bump the tabulate number).

The last optimization is not checking every column: it is enough to
start in the previous column of the left-most edge, and then move right
from there.